Updated Game Scenes and GameStates

+++ source/AA2_02_Arkanoid/Ball/Ball.h
@@ -6,15 +6,15 @@
 class Ball {
 
 public:
-	Ball(const Vector2D<int>& position, const float& moveSpeed);
+	Ball(const Vector2D<float>& position, const float& moveSpeed);
 	~Ball();
 	void InitSprite(SDL_Renderer* renderer);
-	void Move(const Vector2D<int>& direction, const float& elapsedTime);
+	void Move(const Vector2D<float>& direction, const float& elapsedTime);
 	void Draw() const;
 
 private:
 	Image* _sprite;
-	Vector2D<int> _position;
+	Vector2D<float> _position;
 	Vector2D<int> _size;
 	float _moveDistance;
 	float _moveSpeed;
