+++ source/AA2_02_Arkanoid/Player/Platform.h
@@ -8,13 +8,13 @@ class Platform {
 public:
 	Platform();
 	~Platform();
-	void Init(SDL_Renderer* renderer, const Vector2D<int>& destinationStart, const Vector2D<int>& destinationSize, const int& moveSpeed);
+	void Init(SDL_Renderer* renderer, const Vector2D<float>& destinationStart, const Vector2D<int>& destinationSize, const int& moveSpeed);
 	void Draw() const;
-	void Move(const Vector2D<int>& direction, const float& elapsedTime);
+	void Move(const Vector2D<float>& direction, const float& elapsedTime);
 
 private:
 	Image* _sprite;
-	Vector2D<int> _position;
+	Vector2D<float> _position;
 	Vector2D<int> _size;
 	float _moveDistance;
 	float _moveSpeed;
